delete fixed

+++ index.js
@@ -29,7 +29,7 @@ server.get('api/posts/:id', (req, res)=>{
             if(post){
                 res.status(200).json(post);
             }else {
-                res.status(404).json({message: 'user not found'})
+                res.status(404).json({message: 'post not found'})
             }
         })
         .cath(error=>{
@@ -54,15 +54,26 @@ server.post('/api/posts', async (req, res)=>{
     }
 })
 server.delete('/api/posts/:id', (req, res)=>{
-    if(findById(req.params.id)===null){
-        res.status(404).json({message:'The post with the specific ID does not exist'})
-    }
-    db.remove(req.params.id)
+    db.findById(req.params.id)
+    .then(post=>{
+        if(post){
+            db.remove(req.params.id)
         .then(count=>{
             res.status(200).json(count);
         })
         .catch(error =>{
             res.status(500).json({message:'error deleting user'})
         })
+        }
+        else{
+            res.status(404).json({message: 'post not found'}) 
+        }
+    })
+    .catch(error=>{
+        res
+                .status(500)
+                .json({message: 'cant fetch post'})
+    })
+    
 })
 server.listen(8000,()=>console.log('API Running on port 8000') )
